Move Budget class to summary package
