+++ data/budget.txt
@@ -1,1 +1,1 @@
-10.1
+5000.0
