+++ src/main/java/moneytracker/command/AddIncomeCommand.java
@@ -1,6 +1,6 @@
 package moneytracker.command;
 
-import moneytracker.budget.Budget;
+import moneytracker.summary.Budget;
 import moneytracker.exception.MoneyTrackerException;
 import moneytracker.parser.Parser;
 import moneytracker.storage.Storage;
